Added a chunk builder system.
What is this? good question. This is an easy way to add multiple parts chunks to the playable level. The plan is to include a grid based editor that is both easy to use  and..that was it. This is something I've never done before. (Second time working with a custom inspector, and my first time trying to make a level editor..) What I've got right now is a bit of the custom inspector, and being able to save the chunks and edit them. Now all that's left is creating some assets to use and include them inside of that. I could've gone with making my own level format but I feel that a prefab would work just fine considering the time constraints that I have to finish this project.